Dnd Root Element Prop (#33)
The Tree component now has a dndRootElement prop that accepts null, undefined, or an HTML Node. If this prop exists, the drag and drop event handlers that are required by react-dnd will be scoped to that element. Any other native drag and drop listeners will work if they are not contained by the dndRootElement.